moved CouchDB and Keycloak code to services
