+++ src/couchdb/couchdb-admin.controller.ts
@@ -1,20 +1,18 @@
 import { Body, Controller, Get, Post, Query } from '@nestjs/common';
 import { BulkUpdateDto } from './bulk-update.dto';
 import { HttpService } from '@nestjs/axios';
-import { catchError, firstValueFrom, map } from 'rxjs';
+import { firstValueFrom, map } from 'rxjs';
 import { ApiOperation } from '@nestjs/swagger';
 import * as credentials from 'src/assets/credentials.json';
-import { ConfigService } from '@nestjs/config';
+import { CouchdbService } from './couchdb.service';
+import { KeycloakService } from './keycloak.service';
 
 @Controller('couchdb-admin')
 export class CouchdbAdminController {
-  private keycloakPassword = this.configService.get('KEYCLOAK_ADMIN_PASSWORD');
-  private keycloakUrl = this.configService.get('KEYCLOAK_URL');
-  private domain = this.configService.get('DOMAIN');
-
   constructor(
     private http: HttpService,
-    private configService: ConfigService,
+    private couchdbService: CouchdbService,
+    private keycloakService: KeycloakService,
   ) {}
 
   @ApiOperation({
@@ -50,7 +48,11 @@ export class CouchdbAdminController {
     const editedOrgs: string[] = [];
     for (const cred of credentials) {
       const file = `/app/Config:CONFIG_ENTITY`;
-      const config = await this.getDataFromDB(cred.name, file, cred.password);
+      const config = await this.couchdbService.getDataFromDB(
+        cred.name,
+        file,
+        cred.password,
+      );
 
       const configString = JSON.stringify(config);
       const regex = new RegExp(searchString, 'g');
@@ -75,7 +77,11 @@ export class CouchdbAdminController {
     // Update `credentials.json` using the `collect_credentials.sh` script on the server
     for (const cred of credentials) {
       const file = `/app/Config:CONFIG_ENTITY`;
-      const config = await this.getDataFromDB(cred.name, file, cred.password);
+      const config = await this.couchdbService.getDataFromDB(
+        cred.name,
+        file,
+        cred.password,
+      );
 
       const configString = JSON.stringify(config);
       const regex = new RegExp(searchString, 'g');
@@ -83,7 +89,7 @@ export class CouchdbAdminController {
       if (configString.match(regex)) {
         const replaced = configString.replace(regex, replaceString);
         editedOrgs.push(cred.name);
-        await this.putDataToDB(
+        await this.couchdbService.sendDataToDB(
           cred.name,
           file,
           JSON.parse(replaced),
@@ -92,36 +98,6 @@ export class CouchdbAdminController {
       }
     }
     return editedOrgs;
-  }
-
-  private getDataFromDB(org: string, path: string, password: string) {
-    const auth = { username: 'admin', password };
-    const url = `https://${org}.${this.domain}/db`;
-    return firstValueFrom(
-      this.http.get(`${url}/couchdb${path}`, { auth }).pipe(
-        catchError(() => this.http.get(`${url}${path}`, { auth })),
-        map((res) => res.data.rows ?? res.data),
-      ),
-    );
-  }
-
-  private putDataToDB(
-    org: string,
-    path: string,
-    data,
-    password: string,
-    method = this.http.put,
-  ) {
-    const auth = { username: 'admin', password };
-    const url = `https://${org}.${this.domain}/db`;
-    return firstValueFrom(
-      method.call(this.http, `${url}/couchdb${path}`, data, { auth }).pipe(
-        catchError(() =>
-          method.call(this.http, `${url}${path}`, data, { auth }),
-        ),
-        map((res: any) => res.data?.docs),
-      ),
-    );
   }
 
   @ApiOperation({
@@ -135,22 +111,24 @@ export class CouchdbAdminController {
       childrenActive: number;
       users: number;
     }[] = [];
-    const token = await this.getKeycloakToken();
+    const token = await this.keycloakService.getKeycloakToken();
     const allUsers =
       '/_users/_all_docs?startkey="org.couchdb.user:"&endkey="org.couchdb.user:\uffff"';
     const allChildren =
       '/app/_all_docs?startkey="Child:"&endkey="Child:\uffff"';
     const activeChildren = '/app/_find';
     for (const cred of credentials) {
-      const users = await this.getUsersFromKeycloak(cred.name, token).catch(
-        () => this.getDataFromDB(cred.name, allUsers, cred.password),
-      );
-      const children = await this.getDataFromDB(
+      const users = await this.keycloakService
+        .getUsersFromKeycloak(cred.name, token)
+        .catch(() =>
+          this.couchdbService.getDataFromDB(cred.name, allUsers, cred.password),
+        );
+      const children = await this.couchdbService.getDataFromDB(
         cred.name,
         allChildren,
         cred.password,
       );
-      const active: any = await this.putDataToDB(
+      const active: any = await this.couchdbService.sendDataToDB(
         cred.name,
         activeChildren,
         activeChildrenFilter,
@@ -165,33 +143,6 @@ export class CouchdbAdminController {
       });
     }
     return stats;
-  }
-
-  private getKeycloakToken(): Promise<string> {
-    const body = new URLSearchParams();
-    body.set('username', 'admin');
-    body.set('password', this.keycloakPassword);
-    body.set('grant_type', 'password');
-    body.set('client_id', 'admin-cli');
-    return firstValueFrom(
-      this.http
-        .post<{ access_token: string }>(
-          `${this.keycloakUrl}/realms/master/protocol/openid-connect/token`,
-          body.toString(),
-        )
-        .pipe(map((res) => res.data.access_token)),
-    );
-  }
-
-  private getUsersFromKeycloak(org: string, token: string) {
-    return firstValueFrom(
-      this.http
-        .get<{ access_token: string }>(
-          `${this.keycloakUrl}/admin/realms/${org}/users?enabled=true`,
-          { headers: { Authorization: `Bearer ${token}` } },
-        )
-        .pipe(map((res) => res.data)),
-    );
   }
 }
